rest info kado, dan kategori info kado selesai

+++ src/main/java/com/example/carikado/repository/GiftInfoRepository.java
@@ -1,9 +1,6 @@
 package com.example.carikado.repository;
 
 import com.example.carikado.model.GiftInfo;
-import com.example.carikado.model.GiftInfoCategory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
@@ -11,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GiftInfoRepository extends JpaRepository<GiftInfo, Long> {
+public interface GiftInfoRepository extends JpaRepository<GiftInfo, Integer> {
 
     @Query("select gi from GiftInfo gi order by gi.createdAt")
     public List<GiftInfo> findAllWithSort();
